unzip files

+++ volcsarvatory_monitoring/src/main.py
@@ -4,6 +4,7 @@ import json
 import logging
 import os
 import subprocess
+import zipfile
 from pathlib import Path
 
 import asf_search as asf
@@ -196,17 +197,21 @@ def transfer_file(product: str) -> None:
     """
     private_key_str = get_secret('SSH_KEY')
     key_file_path = Path('/tmp/ssh_key.pem')
-    key_file_path.parent.mkdir(parents=True)
-    with key_file_path.open('w') as f:
-        f.write(private_key_str)
-    # Change permissions as required by SSH (read-only for the owner)
-    key_file_path.chmod(0o400)
+    if not key_file_path.parent.exists():
+        key_file_path.parent.mkdir(parents=True)
+        with key_file_path.open('w') as f:
+            f.write(private_key_str)
+        # Change permissions as required by SSH (read-only for the owner)
+        key_file_path.chmod(0o400)
 
     bucket_name = os.environ.get('PUBLISH_BUCKET')
     product_path = Path(product)
     product_path.parent.mkdir(parents=True)
     s3 = boto3.client('s3')
     s3.download_file(bucket_name, product, product)
+    with zipfile.ZipFile(product, mode="r") as archive:
+        archivo.extractall(str(product_path.parent))
+    product_path.unlink()
     ssh_opts = [
         '-i',
         '/tmp/ssh_key.pem',
@@ -216,7 +221,7 @@ def transfer_file(product: str) -> None:
     remote_base = '/shared/data/geodesy/overlay-test'
     dest = f'{remote}:{remote_base}/insar/'
     subprocess.run(['scp', *ssh_opts, str(product_path.parent.parent), dest], check=True)
-    product_path.unlink()
+    subprocess.run(['rm', '-rf', str(product_path.parent.parent)], check=True)
 
 
 def lambda_bucket_handler(event: dict, context: object) -> dict:
